rename AFfiliateLink file
